move extentions in their respective files
